add active frame reducer tests
